CHG: Adjusting the color scheme throughout the GUI
That is, the population is green, visitor demand is bright purple,
walking is bright green, the number of employment is royal blue, and
congestion is dark purple(There is a possibility that dark orange would
be better.).

Also, the use of white as a chart color has often been seen as a problem
in the past.