ADD: A function that allows the fare to be changed by the vehicle alone
Until now, there was only the function to change all convoy vehicles at once.
For example, the acomo rank 4 in convoy is changed to the fare rank 3 at once.
Therefore, when dividing the fare with the same acomo rank (the vehicle and comfort level may be different even if the rank is the same), it was necessary to have a troublesome method of changing and rearranging each car with depot.